Fix errors in serializers and deserializers
Fixed issues with deserializing missing multipart table features,
flows and unsigned longs.
Fixed issues when serializing Ipv4 and Ipv6 source and destination
match entries.
Added additional safety checks to match entry deserializers.

See also: bug 7139, bug 7141

+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/protocol/serialization/actions/SetQueueActionSerializer.java
@@ -19,7 +19,7 @@ public class SetQueueActionSerializer extends AbstractActionSerializer {
     public void serialize(Action action, ByteBuf outBuffer) {
         super.serialize(action, outBuffer);
         final SetQueueAction setQueueAction = SetQueueActionCase.class.cast(action).getSetQueueAction();
-        outBuffer.writeByte(setQueueAction.getQueueId().intValue());
+        outBuffer.writeInt(setQueueAction.getQueueId().intValue());
     }
 
     @Override
